# aethera.client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Aethera.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Characters": {
      "get": {
        "tags": [
          "Characters"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterPreview"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterPreview"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterPreview"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Characters"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Characters/{id}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterDetailDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterDetailDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CharacterDetailDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Characters/equip": {
      "put": {
        "tags": [
          "Characters"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipItemCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipItemCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EquipItemCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Characters/{id}/dynasty": {
      "put": {
        "tags": [
          "Characters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Characters/{id}/parents": {
      "put": {
        "tags": [
          "Characters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetParentsCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetParentsCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetParentsCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Characters/AddTranslation": {
      "post": {
        "tags": [
          "Characters"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTranslationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTranslationCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddTranslationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Dynasties": {
      "post": {
        "tags": [
          "Dynasties"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDynastyCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDynastyCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDynastyCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Dynasties"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DynastyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DynastyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DynastyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Dynasties/{id}": {
      "get": {
        "tags": [
          "Dynasties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DynastyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DynastyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DynastyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Items": {
      "post": {
        "tags": [
          "Items"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddTranslationCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "feats": {
            "type": "string",
            "nullable": true
          },
          "backstory": {
            "type": "string",
            "nullable": true
          },
          "personality": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Alignment": {
        "enum": [
          "LawfulGood",
          "NeutralGood",
          "ChaoticGood",
          "LawfulNeutral",
          "TrueNeutral",
          "ChaoticNeutral",
          "LawfulEvil",
          "NeutralEvil",
          "ChaoticEvil"
        ],
        "type": "string"
      },
      "ArmorType": {
        "enum": [
          "Light",
          "Medium",
          "Heavy"
        ],
        "type": "string"
      },
      "Art": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "AttributeScore": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "modifier": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CharacterClass": {
        "enum": [
          "Fighter",
          "Wizard",
          "Rogue",
          "Cleric",
          "Ranger",
          "Paladin",
          "Bard",
          "Warlock",
          "Sorcerer",
          "Druid",
          "Monk",
          "Barbarian"
        ],
        "type": "string"
      },
      "CharacterDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "characterClass": {
            "$ref": "#/components/schemas/CharacterClass"
          },
          "strengthScore": {
            "$ref": "#/components/schemas/AttributeScore"
          },
          "dexterityScore": {
            "$ref": "#/components/schemas/AttributeScore"
          },
          "constitutionScore": {
            "$ref": "#/components/schemas/AttributeScore"
          },
          "intelligenceScore": {
            "$ref": "#/components/schemas/AttributeScore"
          },
          "wisdomScore": {
            "$ref": "#/components/schemas/AttributeScore"
          },
          "charismaScore": {
            "$ref": "#/components/schemas/AttributeScore"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "experiencePoints": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "size": {
            "$ref": "#/components/schemas/Size"
          },
          "alignment": {
            "$ref": "#/components/schemas/Alignment"
          },
          "art": {
            "$ref": "#/components/schemas/Art"
          },
          "dynastyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "fatherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "motherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "hometownId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "proficiencyBonus": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "initiative": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "armorClass": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "speed": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "passivePerception": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "hp": {
            "$ref": "#/components/schemas/HitPoints"
          },
          "hitDice": {
            "$ref": "#/components/schemas/Dice"
          },
          "deathSaveSuccesses": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "deathSaveFailures": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "skillProficiencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "nullable": true
          },
          "languageProficiencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "feats": {
            "type": "string",
            "nullable": true
          },
          "heroicInspirationCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "backstory": {
            "type": "string",
            "nullable": true
          },
          "personality": {
            "type": "string",
            "nullable": true
          },
          "relatives": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelativeDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CharacterPreview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "class": {
            "$ref": "#/components/schemas/CharacterClass"
          },
          "hp": {
            "$ref": "#/components/schemas/HitPoints"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "art": {
            "$ref": "#/components/schemas/Art"
          },
          "dynastyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateCharacterCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "class": {
            "$ref": "#/components/schemas/CharacterClass"
          }
        },
        "additionalProperties": false
      },
      "CreateDynastyCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateItemCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "type": {
            "$ref": "#/components/schemas/ItemType"
          },
          "damageExpression": {
            "$ref": "#/components/schemas/DiceExpression"
          },
          "armorModifier": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "armorType": {
            "$ref": "#/components/schemas/ArmorType"
          }
        },
        "additionalProperties": false
      },
...
      "Dice": {
        "type": "object",
        "properties": {
          "sides": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DiceExpression": {
        "type": "object",
        "properties": {
          "constantModifier": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "diceCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "diceSides": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DynastyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "art": {
            "$ref": "#/components/schemas/Art"
          },
          "motto": {
            "type": "string",
            "nullable": true
          },
          "establishedYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "power": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/DynastyStatus"
          }
        },
        "additionalProperties": false
      },
      "DynastyStatus": {
        "enum": [
          "Ruling",
          "Fallen",
          "Vassal"
        ],
        "type": "string"
      },
...
      "EntityState": {
        "enum": [
          "Updated",
          "Disabled",
          "Deleted"
        ],
        "type": "string"
      },
      "EquipItemCommand": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "HitPoints": {
        "type": "object",
        "properties": {
          "max": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "current": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "temp": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "entityState": {
            "$ref": "#/components/schemas/EntityState"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "art": {
            "$ref": "#/components/schemas/Art"
          }
        },
        "additionalProperties": false
      },
      "ItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "art": {
            "$ref": "#/components/schemas/Art"
          }
        },
        "additionalProperties": false
      },
      "ItemType": {
        "enum": [
          "Item",
          "Weapon",
          "Armor",
          "Equipment"
        ],
        "type": "string"
      },
      "Language": {
        "enum": [
          "Common",
          "Dwarvish",
          "Elvish",
          "Giant",
          "Gnomish",
          "Goblin",
          "Halfling",
          "Orcish",
          "Abyssal",
          "Celestial",
          "Draconic",
          "DeepSpeech",
          "Infernal",
          "Primordial",
          "Sylvan",
          "Undercommon"
        ],
        "type": "string"
      },
      "RelativeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "fatherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "motherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SetParentsCommand": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "format": "uuid"
          },
          "fatherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "motherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Size": {
        "enum": [
          "Tiny",
          "Small",
          "Medium",
          "Large",
          "Huge",
          "Gargantuan"
        ],
        "type": "string"
      },
      "Skill": {
        "enum": [
          "Acrobatics",
          "AnimalHandling",
          "Arcana",
          "Athletics",
          "Deception",
          "History",
          "Insight",
          "Intimidation",
          "Investigation",
          "Medicine",
          "Nature",
          "Perception",
          "Performance",
          "Persuasion",
          "Religion",
          "SleightOfHand",
          "Stealth",
          "Survival"
        ],
        "type": "string"
      },
      "Species": {
        "enum": [
          "Human",
          "Elf",
          "Dwarf",
          "Halfling",
          "Orc",
          "Gnome",
          "Tiefling",
          "Dragonborn"
        ],
        "type": "string"
      }
    }
  },
...
}